move fake servers to testonly/fake
