+++ testonly/fake/fake_map.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testonly
+package fake
 
 import (
 	"context"
